refactor(utils): refactor validator && readability
- InputValidator: 가격 단위 책임을 input 으로 변경
- LottoPrize: else 문 없애고 리턴
- LottoValidator: 중복 로직 수정
